+++ src/main/java/xyz/theprogramsrc/supercoreapi/spigot/gui/objets/GuiModel.java
@@ -2,11 +2,13 @@ package xyz.theprogramsrc.supercoreapi.spigot.gui.objets;
 
 import java.util.LinkedHashMap;
 
+import xyz.theprogramsrc.supercoreapi.spigot.SpigotModule;
+
 /**
  * Representation of the Gui Model
  * @since 5.2.0
  */
-public class GuiModel {
+public class GuiModel extends SpigotModule{
 
     private GuiTitle title;
     private GuiRows rows;
@@ -46,6 +48,9 @@ public class GuiModel {
         return buttons;
     }
 
+    /**
+     * Clear the buttons from the Gui
+     */
     public void clearButtons(){
         this.buttons.clear();
     }
@@ -72,6 +77,47 @@ public class GuiModel {
         }
     }
 
+    /**
+     * Fills all the empty slots with the given entry
+     * @param entry The entry of the button to add to all the empty slots
+     * @since 5.2.1
+     */
+    public void fillEmptySlotsWithEntry(GuiEntry entry){
+        for(int i = 0; i < this.rows.size; i++){
+            if(!this.buttons.containsKey(i)){
+                this.setButton(i, entry);
+            }
+        }
+    }
+
+    /**
+     * Fills the empty slots with the empty item. 
+     * See {@link xyz.theprogramsrc.supercoreapi.spigot.items.PreloadedItems#emptyItem}
+     * @since 5.2.1
+     */
+    public void fillEmptySlots(){
+        this.fillEmptySlotsWithEntry(new GuiEntry(this.getPreloadedItems().emptyItem()));
+    }
     
-    
+    /**
+     * Fills the specified slots with the specified entry.
+     * @param entry The entry of the button to add to the given slots
+     * @param slots The slots to fill
+     * @since 5.2.1
+     */
+
+    public void fillWithEntry(GuiEntry entry, int[] slots){
+        for(int slot : slots){
+            this.setButton(slot, entry);
+        }
+    }
+
+    /**
+     * Fills the specified slots with an empty item.
+     * @param slots The slots to fill with the empty item
+     * @since 5.2.1
+     */
+    public void fillWithEmptyItem(int[] slots){
+        this.fillWithEntry(new GuiEntry(this.getPreloadedItems().emptyItem()), slots);
+    }
 }